+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 * Lock Mechanism accessories for Omni access controls
 * Temperature & Humidity Sensor accessories for Omni auxiliary sensors ([#5](https://github.com/mantorok1/homebridge-omnilink-platform/issues/5))
+* Show Request/Response in Homebridge logs
 
 ## 1.3.4 (2021-03-07)